修复女皇腺体召唤的虫子在结束时间后不消失的问题

+++ Source/Hediffs/Comp_NoCorpse.cs
@@ -91,7 +91,11 @@ namespace NzFaceLessManMod
             base.CompPostPostRemoved();
             if (Props.deathOnRemove)
             {
-                this.parent.pawn?.Kill(null, null);
+                if (this.parent?.pawn?.Dead == false)
+                {
+                    this.parent.pawn?.Kill(null, null);
+                    this.Notify_PawnDied(null, null);
+                }
             }
         }
     }
